docs site cleanup
