+++ docs/astro.config.mjs
@@ -2,5 +2,9 @@ export default {
   buildOptions: {
     site: 'https://docs.astro.build/',
   },
-  renderers: ['@astrojs/renderer-preact', '@astrojs/renderer-react'],
+  renderers: [
+    // Our main renderer for frontend components
+    '@astrojs/renderer-preact', 
+    // Needed for Algolia search component
+    '@astrojs/renderer-react'],
 };
