+++ docs/src/components/Footer/AvatarList.astro
