+++ docs/src/components/HeadSEO.astro
@@ -1,38 +1,36 @@
 ---
+import {SITE, OPEN_GRAPH} from '../config.ts';
 export interface Props {
   content: any,
   site: any,
   canonicalURL: URL | string,
 };
-const { content = {}, site, canonicalURL } = Astro.props;
-const formattedContentTitle = content.title ? `${content.title} 🚀 ${site.title}` : site.title;
-const imageSrc = content?.image?.src ?? site.image.src;
+const { content = {}, canonicalURL } = Astro.props;
+const formattedContentTitle = content.title ? `${content.title} 🚀 ${SITE.title}` : SITE.title;
+const imageSrc = content?.image?.src ?? OPEN_GRAPH.image.src;
 const canonicalImageSrc = new URL(imageSrc, Astro.site);
-const imageAlt = content?.image?.alt ?? site.image.alt;
+const imageAlt = content?.image?.alt ?? OPEN_GRAPH.image.alt;
 ---
+<!-- Page Metadata -->
+<link rel="canonical" href={canonicalURL}/>
 
 <!-- OpenGraph Tags -->
 <meta property="og:title" content={formattedContentTitle}/>
 <meta property="og:type" content="article"/>
 <meta property="og:url" content={canonicalURL}/>
-<meta property="og:locale" content={content.ogLocale ?? site.ogLocale}/>
+<meta property="og:locale" content={content.ogLocale ?? OPEN_GRAPH.locale}/>
 <meta property="og:image" content={canonicalImageSrc}/>
 <meta property="og:image:alt" content={imageAlt}/>
-<meta property="og:description" content={content.description ? content.description : site.description}/>
-<meta property="og:site_name" content={site.title}/>
-<!-- END OpenGraph Tags -->
+<meta property="og:description" content={content.description ? content.description : SITE.description}/>
+<meta property="og:site_name" content={SITE.title}/>
 
 <!-- Twitter Tags -->
 <meta name="twitter:card" content="summary_large_image"/>
-<meta name="twitter:site" content={site.twitter.site}/>
-<meta name="twitter:creator" content={site.twitter.creator}/>
+<meta name="twitter:site" content={OPEN_GRAPH.twitter}/>
 <meta name="twitter:title" content={formattedContentTitle}/>
-<meta name="twitter:description" content={content.description ? content.description : site.description}/>
+<meta name="twitter:description" content={content.description ? content.description : SITE.description}/>
 <meta name="twitter:image" content={canonicalImageSrc}/>
 <meta name="twitter:image:alt" content={imageAlt}/>
-<!-- END Twitter Tags -->
-
-<link rel="canonical" href={canonicalURL}/>
 
 <!-- 
   TODO: Add json+ld data, maybe https://schema.org/APIReference makes sense?
