+++ docs/src/components/PageContent/PageContent.astro
@@ -1,0 +1,41 @@
+---
+const {content, githubEditUrl} = Astro.props;
+const title = content.title;
+const headers = content.astro.headers;
+import MoreMenu from '../RightSidebar/MoreMenu.astro';
+import TableOfContents from '../RightSidebar/TableOfContents.tsx';
+---
+<style>
+      .content {
+        padding: 0;
+        max-width: 75ch;
+        width: 100%;
+        height: 100%;
+        display: flex;
+        flex-direction: column;
+      }
+      .content > section {
+        margin-bottom: 4rem;
+      }
+      .block {
+        display: block;
+      }
+
+      @media (min-width: 50em) {
+        .sm\:hidden {
+          display: none;
+        }
+      }
+</style>
+<article id="article" class="content">
+    <section class="main-section">
+        <h1 class="content-title" id="overview">{title}</h1>
+        <nav class="block sm:hidden">
+            <TableOfContents client:media="(max-width: 50em)" headers={headers}/>
+        </nav>
+        <slot />
+    </section>
+    <nav class="block sm:hidden">
+        <MoreMenu editHref={githubEditUrl}/>
+    </nav>
+</article>
