--- docs/src/components/Search.css
@@ -1,137 +1,0 @@
-:root {
-  --docsearch-primary-color: var(--theme-accent);
-  --docsearch-logo-color: var(--theme-text);
-}
-
-.search {
-  display: none;
-  position: relative;
-  z-index: 10;
-  flex-grow: 1;
-  grid-area: search;
-  padding-right: 0.7rem;
-  display: flex;
-  max-width: 200px;
-}
-.search > * {
-  flex-grow: 1;
-}
-
-.search > :global(.algolia-autocomplete) {
-  width: 100%;
-}
-
-:global(body.is-nav-open) .search {
-  display: flex;
-}
-
-.search-input {
-  flex-grow: 1;
-  box-sizing: border-box;
-  width: 100%;
-  margin: 0;
-  padding: 0.33em 0.5em;
-  overflow: visible;
-  font-weight: 500;
-  font-size: 1rem;
-  font-family: inherit;
-  line-height: inherit;
-  background-color: var(--theme-divider);
-  border-color: var(--theme-divider);
-  color: var(--theme-text-light);
-  border-style: solid;
-  border-width: 1px;
-  border-radius: 0.25rem;
-  outline: 0;
-  cursor: pointer;
-  transition-timing-function: ease-out;
-  transition-duration: 0.2s;
-  transition-property: border-color, color;
-  -webkit-font-smoothing: antialiased;
-}
-.no-underline {
-  text-decoration: none;
-}
-.search-input:hover {
-  color: var(--theme-text);
-}
-.search-input:hover::placeholder {
-  color: var(--theme-text-light);
-}
-.search-input::placeholder {
-  color: var(--theme-text-light);
-}
-
-.search-hint {
-  position: absolute;
-  top: 7px;
-  right: 19px;
-  padding: 3px 5px;
-  display: none;
-  display: none;
-  align-items: center;
-  justify-content: center;
-  letter-spacing: 0.125em;
-  font-size: 13px;
-  font-family: var(--font-mono);
-  pointer-events: none;
-  border-color: var(--theme-text-lighter);
-  color: var(--theme-text-light);
-  border-style: solid;
-  border-width: 1px;
-  border-radius: 0.25rem;
-  line-height: 14px;
-}
-
-.select-language {
-  flex-grow: 1;
-  width: 32px;
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0.33em 0.5em;
-  overflow: visible;
-  font-weight: 500;
-  font-size: 1rem;
-  font-family: inherit;
-  line-height: inherit;
-  background-color: var(--theme-bg);
-  border-color: var(--theme-text-lighter);
-  color: var(--theme-text-light);
-  border-style: solid;
-  border-width: 1px;
-  border-radius: 0.25rem;
-  outline: 0;
-  cursor: pointer;
-  transition-timing-function: ease-out;
-  transition-duration: 0.2s;
-  transition-property: border-color, color;
-  -webkit-font-smoothing: antialiased;
-  padding-left: 30px;
-  padding-right: 1rem;    
-}
-.language-select:hover,
-.language-select:hover .select-language {
-  color: var(--theme-text);
-  border-color: var(--theme-text-light);
-}
-.language-select {
-  color: var(--theme-text-light);
-  position: relative;
-}
-.language-select > svg {
-    position: absolute;
-    top: 6px;
-    left: 10px;
-}
-
-@media (min-width: 50em) {
-  .search {
-    max-width: 400px;
-  }
-  .search-hint {
-    display: flex;
-  }
-  .select-language {
-    width: 100%;
-  }
-}
