+++ docs/src/pages/blog/island-architecture.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: Island Architecture
 draft: true
 ---
