+++ docs/src/pages/comparing-astro-vs-other-tools.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: Comparing Astro
 ---
