+++ docs/src/pages/core-concepts/astro-components.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: Astro Components
 ---
