+++ docs/src/pages/core-concepts/astro-pages.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: Pages
 ---
