+++ docs/src/pages/core-concepts/component-hydration.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: Partial Hydration in Astro
 ---
