+++ docs/src/pages/core-concepts/layouts.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: Layouts
 ---
