+++ docs/src/pages/core-concepts/project-structure.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: Project Structure
 ---
