+++ docs/src/pages/guides/data-fetching.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: Data Fetching
 ---
