+++ docs/src/pages/guides/deploy.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: Deploy a Website
 ---
