+++ docs/src/pages/guides/markdown-content.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: Markdown
 ---
