+++ docs/src/pages/guides/publish-to-npm.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: Publish a Component to NPM
 ---
