+++ docs/src/pages/guides/styling.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: Styling & CSS
 ---
