+++ docs/src/pages/index.astro
@@ -1,9 +1,16 @@
 ---
-import Layout from '../layouts/Main.astro';
+import Layout from '../layouts/MainLayout.astro';
 ---
 
-<Layout centered>
-  <h1>
-    <code style="white-space: nowrap;">npm init astro</code>
-  </h1>
-</Layout>
+<script>
+  // This is some temporary code for now.
+  const KNOWN_LANGUAGES = ['en', 'nl'];
+  let newLang = (window.navigator.userLanguage || window.navigator.language || 'en').substr(0, 2);
+  if (newLang === 'en') {
+    window.location.pathname = '/getting-started';
+  } else if (KNOWN_LANGUAGES.includes(newLang)) {
+    window.location.pathname = '/' + newLang + '/getting-started';
+  } else {
+  window.location.pathname = '/getting-started';
+  }
+</script>
