+++ docs/src/pages/nl/getting-started.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: Beginnen
 ---
