+++ docs/src/pages/reference/api-reference.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: API Reference
 ---
