+++ docs/src/pages/reference/builtin-components.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: Built-In Components
 ---
