+++ docs/src/pages/reference/configuration-reference.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: Configuration Reference
 ---
