+++ docs/src/pages/reference/renderer-reference.md
@@ -1,5 +1,5 @@
 ---
-layout: ~/layouts/Main.astro
+layout: ~/layouts/MainLayout.astro
 title: UI Renderer Reference
 ---
