+++ docs/vercel.json
@@ -1,10 +1,4 @@
 {
   "cleanUrls": true,
-  "trailingSlash": false,
-  "redirects": [
-    {
-      "source": "/",
-      "destination": "/getting-started"
-    }
-  ]
+  "trailingSlash": false
 }
